+++ imagesmacker/models/fields.py
@@ -7,7 +7,7 @@ from imagesmacker.models.coordinates import XYXY
 
 cells_type: TypeAlias = Sequence[Union["RelativeContainer", "RelativeFieldCell"]]
 directions_type: TypeAlias = Literal["lr", "rl", "tb", "bt"]
-cell_variant_types: TypeAlias = Literal["text", "barcode.code128","barcode.qr"]
+cell_variant_types: TypeAlias = Literal["text", "barcode.code128", "barcode.qr"]
 
 
 class RelativeFieldCell(BaseModel):